pyproject.toml
[tool.poetry]
name = "ska-mid-cbf-fhs-system-tests"
version = "0.0.2"
description = "System tests for H1 and beyond."
authors = [
    "Justin Wamback <justin.wamback@mda.space>",
    "Ben Herriott <ben.herriott@mda.space>",
    "Jason Turner <jason.turner@mda.space>",
    "Theo Lemay <theo.lemay@mda.space>",
]
license = "BSD license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jsonref = "^1.1.0"
pytango = "9.5.0"
requests = { version = "2.28.0", python = ">=3.7,<4" }
ska-mid-cbf-internal-schemas = { version = "0.4.3+dev.c37e717e0", python = ">=3.8,<4.0", source = "gitlab-internal-schemas" }
ska-tango-base = "1.0.0"
ska-tango-testing = "0.7.1"
typing-extensions = "^4.12.2"

[[tool.poetry.source]]
name = "nexus-internal"
url = "https://artefact.skao.int/repository/pypi-internal/simple"

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'

[[tool.poetry.source]]
name = "gitlab-internal-schemas"
url = "https://gitlab.com/api/v4/projects/47018613/packages/pypi/simple"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pylint = "^3.2.3"
pylint-junit = "^0.3.2"
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
pytest-forked = "^1.6.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.11.1"
pytest-ordering = "^0.6"
pytest-pycodestyle = "2.2.0"
pytest-pydocstyle = "2.2.0"
pytest-pylint = "^0.21.0"
pytest-rabbitmq = "^3.1.0"
pytest-repeat = "^0.9.1"
pytest-timeout = "^2.1.0"
pytest-xdist = "^3.3.1"
python-dotenv = "^0.21.0"
pyyaml = "^6.0.1"
types-requests = "^2.28.11.2"

[tool.poetry.group.docs.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"

[tool.pytest.ini_options]
filterwarnings = [
  "ignore::_pytest.warning_types.PytestUnknownMarkWarning",
  "ignore::tango.PyTangoUserWarning",
  "ignore::pytest_cov.plugin.CovDisabledWarning"
]
addopts = "--show-capture=no"
log_cli = 1

[build-system]
requires = ["poetry>=1.8", "poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
